use new multiplication in findperiod too
