+++ src/test/java/org/redfx/strange/test/Arithmetic4Tests.java
@@ -427,91 +427,6 @@ public class Arithmetic4Tests extends BaseGateTests {
     }
   
     @Test
-    public void testexpmodH37() {
-        //  H { 1 x 3 ^ A mod 7}
-        int mod = 7;
-        int a = 3;
-        int length = (int) Math.ceil(Math.log(mod) / Math.log(2));
-        int offset = 1 ;
-        int dim0 = 1 << offset;
-        assertEquals(dim0, 2);
-        Program p = new Program(2 * length + 3 + offset);
-        Step prep = new Step();
-
-        prep.addGate(new X(3));
-        
-        Step prepAnc = new Step(new X(length + 1 + offset));
-        p.addStep(prep);
-        p.addStep(prepAnc);
-        for (int i = offset - 1; i > - 1; i--) {
-            int m = 1;
-            for (int j = 0; j < 1 << i; j++) {
-                m = m * a % mod;
-            }
-//            System.err.println("Create MulModulus, i = "+i+", m = "+m+", a = "+a+", MOD = "+mod);
-            MulModulus mul = new MulModulus(length, 2 * length, m, mod);
-            ControlledBlockGate cbg = new ControlledBlockGate(mul, offset, i);
-            p.addStep(new Step(cbg));
-        }
-        p.addStep(new Step(new InvFourier(offset, 0)));
-        Result result = runProgram(p);
-        Complex[] probs = result.getProbability();
-        Qubit[] q = result.getQubits();
-        int answer = 0;
-        double[] amps = new double[dim0];
-        for (int i = 0; i < probs.length; i++) {
-            amps[i%dim0] = amps[i%dim0] + probs[i].abssqr();
-        }
-        for (int i = 0; i < dim0; i ++) {
-//            System.err.println("AMP["+i+"] = "+amps[i]);
-            assertEquals(amps[i], .5, 0.001);
-        }
-    }
-    
-    @Test
-    public void testexpmodHH37() {
-        //  H { 1 x 3 ^ A mod 7}
-        int mod = 7;
-        int a = 3;
-        int length = (int) Math.ceil(Math.log(mod) / Math.log(2));
-        int offset = 2 ;
-        int dim0 = 1 << offset;
-        assertEquals(dim0, 4);
-        Program p = new Program(2 * length + 3 + offset);
-        Step prep = new Step();
-        for (int i = 0; i < offset; i++) {
-            prep.addGate(new Hadamard(i));
-        }
-        
-        Step prepAnc = new Step(new X(length + 1 + offset));
-        p.addStep(prep);
-        p.addStep(prepAnc);
-        for (int i = offset - 1; i > - 1; i--) {
-            int m = 1;
-            for (int j = 0; j < 1 << i; j++) {
-                m = m * a % mod;
-            }
-//            System.err.println("Create MulModulus, i = "+i+", m = "+m+", a = "+a+", MOD = "+mod);
-            MulModulus mul = new MulModulus(length, 2 * length, m, mod);
-            ControlledBlockGate cbg = new ControlledBlockGate(mul, offset, i);
-            p.addStep(new Step(cbg));
-        }
-        p.addStep(new Step(new InvFourier(offset, 0)));
-        Result result = runProgram(p);
-        Complex[] probs = result.getProbability();
-        Qubit[] q = result.getQubits();
-        int answer = 0;
-        double[] amps = new double[dim0];
-        for (int i = 0; i < probs.length; i++) {
-            amps[i%dim0] = amps[i%dim0] + probs[i].abssqr();
-        }
-        for (int i = 0; i < dim0; i ++) {
-         //   System.err.println("AMP["+i+"] = "+amps[i]);
-            assertEquals(amps[i], .25, 0.001);
-        }
-    }
-    
-    @Test
     public void testMinOffset() {
         int offset = 2;
         int n = 2;
